Server/snpm_server/requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
pyodbc==4.0.39
html2text==2020.1.16
pyjwt==2.6.0
cryptography==36.0.2
flask-sqlalchemy==3.0.3
flask-swagger-ui==4.11.1
pytest==7.3.1
requests==2.27.1